requirements.txt
# Core Django
Django==5.0.1
django-environ==0.11.2

# Django REST Framework
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0

# Database
psycopg2-binary==2.9.9

# Caching & Session
redis==5.0.1
django-redis==5.4.0
hiredis==2.3.2

# Task Queue
celery==5.3.4
kombu==5.3.4
amqp==5.2.0

# File Storage
Pillow==10.2.0
boto3==1.34.25
django-storages==1.14.2

# API Utilities
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7
drf-nested-routers==0.95.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Payment Processing
stripe==7.10.0

# Security
cryptography==42.0.0
PyJWT==2.8.0

# Web Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Logging
sentry-sdk==1.39.2
python-json-logger==2.0.7

# Utilities
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0

# Email
django-anymail==10.2